requirements.txt
jupyter
matplotlib
numpy
torchsummary
tensorboard
pandas
pydicom
setuptools
scikit-learn
monai
nibabel
opencv-python
scikit-image